Created a template for TextPanel so it will be easier to make global changes.
Turned up a bug that was very hard to find - changing a LayoutOptions was not
propagating to its TextNode; added IsValid() and CopyContentsFrom() to make
sure the TextNode is added as an observer.

Added some code to help debug - templates are now indicated as such when
writing out and are automatically disabled for traversal.