Changed template syntax in Parser: now uses TEMPLATE and INSTANCE so that type names are not associated automatically as templates.
This was a problem when trying to create a template from an SG::Node - it made
all subsequent Nodes into instances of this template. Yikes! Now templates are
keyed by the name of the template object, not its type. This required adding
the INSTANCE syntax to be able to parse the template name.