Fixed TranslationTool slider issue.
Added a kSearch flag to SG::Node to enable/disable searching. Added tests for
it. Use it to skip searching through Tools definitions so that the actual Tool
in the scene is found instead when getting a path to a Widget. Now the correct
path and transformation matrix are used.