Changed TriMesh validation to try to repair problems if possible, updating the mesh in a Model. This makes sure the validation code matches the actual mesh.
However, it turns out that CGAL repair of self-intersections does not actually
retain the overall geometry of the mesh - it throws away a lot of vertices and
faces. So it it useless for this app.

So, #if'ed out the code that does the repair until a time when CGAL works
better.