Implemented translation snapping for TwistTool.
This required simplifying the TwistTool graph structure, which then required a
way to change the plane for Slider2DWidget. The Slider2DWidget now takes an
enum specifying the principal plane for interaction. Added a way to set up the
plane at the start of a drag so that it is useful and consistent across the
drag.