Large refactoring to enable different objects to handle scroll-wheel interaction. Added an IScrollable interface that the MainHandler uses to determine who to pass valuator events to. StageWidget, Inspector, and TreePanel all implement this. This means that the mouse cursor now needs to be over a scrollable object when using the scroll wheel. GLFWWindowSystem now adds the cursor position to the scroll events so the object correlation works.
While refactoring, added an ITargetable interface instead of having
CanReceiveTarget() in the base Widget class. Model and StageWidget implement
this interface.

Since the new interfaces are in the Widgets subdirectory, moved ITouchable
there as well for consistency.

Also changed the Border node to turn off depth testing rather than having it
offset in Z; the small constant offset did not work well in the TreePanel,
which is too far away.